requirements.txt
requests>=2.28.1
numpy>=1.23.3
lxml>=4.9.1
openpyxl >=3.1.2
jinja2 >=3.1.2
python-dotenv>=1.0.0
xlsxwriter>=3.2.0
pandas>=1.5.0